check for file.exists

+++ R/leeds_2018.R
@@ -7,8 +7,12 @@ location = "Leeds"
 year = 2018
 
 # Get OSM Data
-osm = download_osm(paste0(location,", UK"))
-saveRDS(osm, paste0("data/osm_",location,"_raw.Rds"))
+if(file.exists(paste0("data/osm_",location,"_raw.Rds"))){
+  osm = readRDS(paste0("data/osm_",location,"_raw.Rds"))
+} else {
+  osm = download_osm(paste0(location,", UK"))
+  saveRDS(osm, paste0("data/osm_",location,"_raw.Rds"))
+}
 
 # Get Traffic Counts
 traffic_leeds_2018 = prepare_traffic_data("data/dft_aadf_local_authority_id_63.csv", year = year)
